Add STTManager
